Fix style in His_Order and refactor Fragments
