+++ app/src/main/java/com/unipet7/mcommerce/fragments/Fragment_Empty_Notification.java
@@ -1,4 +1,4 @@
-package com.unipet7.mcommerce.fragments.onboarding;
+package com.unipet7.mcommerce.fragments;
 
 import android.os.Bundle;
 
@@ -9,19 +9,13 @@ import android.view.View;
 import android.view.ViewGroup;
 
 import com.unipet7.mcommerce.R;
-import com.unipet7.mcommerce.adapters.HistoryOrderAdapter;
-import com.unipet7.mcommerce.databinding.FragmentConfirmationOrderBinding;
-import com.unipet7.mcommerce.models.HistoryOrders;
-
-import java.util.ArrayList;
-import java.util.List;
 
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link HistoryOrder_Confirm#newInstance} factory method to
+ * Use the {@link Fragment_Empty_Notification#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class HistoryOrder_Confirm extends Fragment {
+public class Fragment_Empty_Notification extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -31,10 +25,8 @@ public class HistoryOrder_Confirm extends Fragment {
     // TODO: Rename and change types of parameters
     private String mParam1;
     private String mParam2;
-    FragmentConfirmationOrderBinding binding;
-    HistoryOrderAdapter HistoryAdapter;
-    ArrayList<HistoryOrders> Historyorders;
-    public HistoryOrder_Confirm() {
+
+    public Fragment_Empty_Notification() {
         // Required empty public constructor
     }
 
@@ -44,11 +36,11 @@ public class HistoryOrder_Confirm extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment ConfirmationOrder.
+     * @return A new instance of fragment Fragment_Empty_Notification.
      */
     // TODO: Rename and change types and number of parameters
-    public static HistoryOrder_Confirm newInstance(String param1, String param2) {
-        HistoryOrder_Confirm fragment = new HistoryOrder_Confirm();
+    public static Fragment_Empty_Notification newInstance(String param1, String param2) {
+        Fragment_Empty_Notification fragment = new Fragment_Empty_Notification();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -69,25 +61,6 @@ public class HistoryOrder_Confirm extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        binding = FragmentConfirmationOrderBinding.inflate(inflater,container,false);
-        loadData();
-        addEvents();
-
-        return binding.getRoot();    }
-
-    private void addEvents() {
-    }
-
-    private void loadData() {
-        HistoryAdapter = new HistoryOrderAdapter(getActivity(),R.layout.layout_confirmorder, initData());
-        binding.lvlConfOrder.setAdapter(HistoryAdapter);
-    }
-
-    private List<HistoryOrders> initData() {
-        Historyorders = new ArrayList<>();
-        Historyorders.add(new HistoryOrders("11/12/2024","1111", "Chờ xác nhận", R.drawable.unipet_app_icon,"Royal Canin Rottweiler Puppy con mèo con đi hia long nhong",
-                "Túi", 1.0,24000.0, 24000.0));
-
-        return Historyorders;
+        return inflater.inflate(R.layout.fragment__empty__notification, container, false);
     }
 }
