+++ app/src/main/java/com/unipet7/mcommerce/fragments/HistoryOrder_Received.java
@@ -1,4 +1,4 @@
-package com.unipet7.mcommerce;
+package com.unipet7.mcommerce.fragments;
 
 import android.os.Bundle;
 
@@ -8,12 +8,20 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.unipet7.mcommerce.R;
+import com.unipet7.mcommerce.adapters.HistoryOrderAdapter;
+import com.unipet7.mcommerce.databinding.FragmentReceivedOrderBinding;
+import com.unipet7.mcommerce.models.HistoryOrders;
+
+import java.util.ArrayList;
+import java.util.List;
+
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link fragment_notification#newInstance} factory method to
+ * Use the {@link HistoryOrder_Received#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class fragment_notification extends Fragment {
+public class HistoryOrder_Received extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -23,8 +31,10 @@ public class fragment_notification extends Fragment {
     // TODO: Rename and change types of parameters
     private String mParam1;
     private String mParam2;
-
-    public fragment_notification() {
+    FragmentReceivedOrderBinding binding;
+    HistoryOrderAdapter HistoryAdapter;
+    ArrayList<HistoryOrders> Historyorders;
+    public HistoryOrder_Received() {
         // Required empty public constructor
     }
 
@@ -34,11 +44,11 @@ public class fragment_notification extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment fragment_notification.
+     * @return A new instance of fragment ReceivedOrder.
      */
     // TODO: Rename and change types and number of parameters
-    public static fragment_notification newInstance(String param1, String param2) {
-        fragment_notification fragment = new fragment_notification();
+    public static HistoryOrder_Received newInstance(String param1, String param2) {
+        HistoryOrder_Received fragment = new HistoryOrder_Received();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -58,7 +68,25 @@ public class fragment_notification extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_notification, container, false);
+        binding = FragmentReceivedOrderBinding.inflate(inflater,container,false);
+        loadData();
+        addEvents();
+
+        return binding.getRoot();    }
+
+    private void addEvents() {
+    }
+
+    private void loadData() {
+        HistoryAdapter = new HistoryOrderAdapter(getActivity(),R.layout.layout_history_order, initData());
+        binding.lvlReceivedOrder.setAdapter(HistoryAdapter);
+    }
+
+    private List<HistoryOrders> initData() {
+        Historyorders = new ArrayList<>();
+        Historyorders.add(new HistoryOrders("11/12/2024","1111", "Đã nhận", R.drawable.unipet_app_icon,"Royal Canin Rottweiler Puppy con mèo con đi hia long nhong",
+                "Túi", 1.0,24000.0, 24000.0));
+
+        return Historyorders;
     }
 }
