+++ app/src/main/java/com/unipet7/mcommerce/fragments/Question.java
@@ -1,16 +1,14 @@
-package com.unipet7.mcommerce.fragments.onboarding;
+package com.unipet7.mcommerce.fragments;
 
 import android.os.Bundle;
 
 import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.LinearLayoutManager;
-import androidx.recyclerview.widget.RecyclerView;
 
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import com.unipet7.mcommerce.R;
 import com.unipet7.mcommerce.adapters.QuestionAdapter;
 import com.unipet7.mcommerce.databinding.FragmentDeliveryOrderBinding;
 import com.unipet7.mcommerce.databinding.FragmentQuestionBinding;
