+++ app/src/main/java/com/unipet7/mcommerce/fragments/fragment_notification.java
@@ -1,4 +1,4 @@
-package com.unipet7.mcommerce;
+package com.unipet7.mcommerce.fragments;
 
 import android.os.Bundle;
 
@@ -8,12 +8,14 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.unipet7.mcommerce.R;
+
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link Fragment_Product#newInstance} factory method to
+ * Use the {@link fragment_notification#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class Fragment_Product extends Fragment {
+public class fragment_notification extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -24,7 +26,7 @@ public class Fragment_Product extends Fragment {
     private String mParam1;
     private String mParam2;
 
-    public Fragment_Product() {
+    public fragment_notification() {
         // Required empty public constructor
     }
 
@@ -34,11 +36,11 @@ public class Fragment_Product extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment Fragment_Product.
+     * @return A new instance of fragment fragment_notification.
      */
     // TODO: Rename and change types and number of parameters
-    public static Fragment_Product newInstance(String param1, String param2) {
-        Fragment_Product fragment = new Fragment_Product();
+    public static fragment_notification newInstance(String param1, String param2) {
+        fragment_notification fragment = new fragment_notification();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -59,6 +61,6 @@ public class Fragment_Product extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment__product, container, false);
+        return inflater.inflate(R.layout.fragment_notification, container, false);
     }
 }
