+++ app/src/main/res/layout/activity_voucher.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".VoucherActivity">
+    tools:context=".activities.VoucherActivity">
 
     <!-- AppBarLayout -->
     <com.google.android.material.appbar.AppBarLayout
