+++ app/src/main/res/layout/fragment__empty__notification.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Fragment_Empty_Notification"
+    tools:context=".fragments.Fragment_Empty_Notification"
     android:paddingHorizontal="20dp"
     android:paddingVertical="20dp">
