+++ app/src/main/res/layout/fragment__product.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Fragment_Product"
+    tools:context=".fragments.Fragment_Product"
     android:padding="20dp">
 
     <LinearLayout
