+++ app/src/main/res/layout/fragment_all_history_order.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".fragments.onboarding.HistoryOrder_All"
+    tools:context=".fragments.HistoryOrder_All"
     android:id="@+id/fragmentAll">
 
     <!-- TODO: Update blank fragment layout -->
