+++ app/src/main/res/layout/fragment_contact_list.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="10dp"
-    tools:context=".fragments.onboarding.ContactList"
+    tools:context=".fragments.ContactList"
     >
     <LinearLayout
         android:layout_width="match_parent"
